pyproject.toml
[tool.poetry]
name = "data-pipelines"
version = "0.1.0"
description = "Data pipelines for the OpenEPI data platform"
authors = ["Aleksander Stangeland <aleksander.stangeland@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.25.2"
dagster = "^1.5.12"
dagster-webserver = "^1.5.12"
bokeh = "^3.3.2"
pyarrow = "^14.0.2"
dask = "^2023.12.0"
distributed = "^2023.12.0"
xarray = "^2023.12.0"
rioxarray = "^0.15.0"
rio-cogeo = "^5.1.0"
geopandas = "^0.14.1"
zarr = "^2.16.1"
geocube = "^0.5.0"
flox = "^0.9.0"
cdsapi = "^0.6.1"
dagster-pyspark = "^0.22.2"
netcdf4 = "^1.6.5"
eccodes = "^1.6.1"
ecmwflibs = "0.5.2"
cfgrib = "^0.9.10.4"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
notebook = "^7.0.6"


[tool.dagster]
module_name = "data_pipelines"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
